check if future completed with error before delete

+++ src/main/java/com/bwarelabs/geyser2cos/Writer.java
@@ -36,8 +36,7 @@ import java.util.concurrent.Executors;
  * The main steps performed by this class are:
  * 1. Identify slot range directories from the input storage path.
  * 2. For each slot range directory, process the subdirectories (slots).
- * 3. For each slot, create sequence files for different categories (entries,
- * blocks, tx, tx_by_addr).
+ * 3. For each slot, create sequence files for different categories (entries, blocks)
  * 4. Write the data to the sequence files.
  * 5. Upload the sequence files to COS asynchronously.
  * 6. Ensure all uploads are completed before finishing the process.
@@ -151,16 +150,15 @@ public class Writer implements AutoCloseable {
                 ResultSerialization.class.getName());
         String range = String.valueOf(slotRangeDir.getFileName());
 
-        try (
-                CustomS3FSDataOutputStream blocksStream = new CustomS3FSDataOutputStream(cosUtils, range, "blocks",
-                        SYNC_TYPE);
-                CustomSequenceFileWriter blocksWriter = new CustomSequenceFileWriter(hadoopConfig, blocksStream);
+        try (Stream<Path> slotDirs = Files.list(slotRangeDir)) {
+            CustomS3FSDataOutputStream blocksStream = new CustomS3FSDataOutputStream(cosUtils, range, "blocks",
+                    SYNC_TYPE);
+            CustomSequenceFileWriter blocksWriter = new CustomSequenceFileWriter(hadoopConfig, blocksStream);
 
-                CustomS3FSDataOutputStream entriesStream = new CustomS3FSDataOutputStream(cosUtils, range, "entries",
-                        SYNC_TYPE);
-                CustomSequenceFileWriter entriesWriter = new CustomSequenceFileWriter(hadoopConfig, entriesStream);
+            CustomS3FSDataOutputStream entriesStream = new CustomS3FSDataOutputStream(cosUtils, range, "entries",
+                    SYNC_TYPE);
+            CustomSequenceFileWriter entriesWriter = new CustomSequenceFileWriter(hadoopConfig, entriesStream);
 
-                Stream<Path> slotDirs = Files.list(slotRangeDir)) {
             slotDirs
                     .filter(Files::isDirectory)
                     .forEach(slotDir -> {
@@ -177,16 +175,24 @@ public class Writer implements AutoCloseable {
             blocksWriter.close();
             entriesWriter.close();
 
-            return CompletableFuture.allOf(
-                    blocksStream.getUploadFuture()).thenRunAsync(() -> {
-                        logger.info("Slot range processed: " + slotRangeDir.getFileName() + ", deleting slot range...");
-                        try {
-                            deleteDirectory(slotRangeDir);
-                            logger.info("Deleted slot range: " + slotRangeDir.getFileName());
-                        } catch (Exception e) {
-                            logger.severe(String.format("Error deleting slot range: %s, %s", slotRangeDir.getFileName(),
-                                    e.getMessage()));
+            return CompletableFuture.allOf(blocksStream.getUploadFuture(), entriesStream.getUploadFuture())
+                    .thenComposeAsync(result -> {
+                        if (blocksStream.getUploadFuture().isCompletedExceptionally() || entriesStream.getUploadFuture().isCompletedExceptionally()) {
+                            logger.severe("One or more uploads failed; skipping deletion for slot range: " + slotRangeDir.getFileName());
+                            logger.severe("Blocks upload future: " + blocksStream.getUploadFuture());
+                            logger.severe("Entries upload future: " + entriesStream.getUploadFuture());
+                            return CompletableFuture.completedFuture(null);
                         }
+
+                        return CompletableFuture.runAsync(() -> {
+                            logger.info("Slot range processed: " + slotRangeDir.getFileName() + ", deleting slot range...");
+                            try {
+                                deleteDirectory(slotRangeDir);
+                                logger.info("Deleted slot range: " + slotRangeDir.getFileName());
+                            } catch (Exception e) {
+                                logger.severe(String.format("Error deleting slot range: %s, %s", slotRangeDir.getFileName(), e.getMessage()));
+                            }
+                        }, executorService);
                     }, executorService);
         } catch (Exception e) {
             logger.severe(
